requirements.main.txt
PyYAML==6.0.2
coloredlogs==15.0.1
mkdocs==1.6.1
mwclient==0.11.0
pypandoc_binary==1.15
python-dotenv==1.0.1
tqdm==4.67.1
type-serialize==1.3.0